+++ presto-main/src/main/java/com/facebook/presto/execution/StageExecutionStateMachine.java
@@ -237,7 +237,7 @@ public class StageExecutionStateMachine
         int runningDrivers = 0;
         int completedDrivers = 0;
 
-        long cumulativeUserMemory = 0;
+        double cumulativeUserMemory = 0;
         long userMemoryReservation = 0;
         long totalMemoryReservation = 0;
